# Bulk append values in Parquet and switch some places to int64